REsloved auth and middleware duplication
