--- backend/middleware/auth.js
@@ -1,176 +1,0 @@
-const jwt = require('jsonwebtoken');
-const User = require('../models/User');
-
-// Generate Access Token
-const generateAccessToken = (payload) => {
-  return jwt.sign(payload, process.env.JWT_SECRET, {
-    expiresIn: process.env.JWT_ACCESS_EXPIRE || '15m'
-  });
-};
-
-// Generate Refresh Token
-const generateRefreshToken = (payload) => {
-  return jwt.sign(payload, process.env.JWT_SECRET, {
-    expiresIn: process.env.JWT_REFRESH_EXPIRE || '7d'
-  });
-};
-
-// Verify Token
-const verifyToken = (token) => {
-  return jwt.verify(token, process.env.JWT_SECRET);
-};
-
-// Authentication Middleware
-const authenticateToken = async (req, res, next) => {
-  try {
-    const authHeader = req.headers.authorization;
-    const token = authHeader && authHeader.split(' ')[1]; // Bearer TOKEN
-
-    if (!token) {
-      return res.status(401).json({
-        success: false,
-        message: 'Access token required'
-      });
-    }
-
-    const decoded = verifyToken(token);
-    const user = await User.findById(decoded.userId).select('-password -refreshTokens');
-    
-    if (!user) {
-      return res.status(401).json({
-        success: false,
-        message: 'User not found'
-      });
-    }
-
-    if (!user.isActive) {
-      return res.status(401).json({
-        success: false,
-        message: 'Account is deactivated'
-      });
-    }
-
-    req.user = user;
-    req.userId = user._id;
-    next();
-  } catch (error) {
-    if (error.name === 'TokenExpiredError') {
-      return res.status(401).json({
-        success: false,
-        message: 'Token expired',
-        code: 'TOKEN_EXPIRED'
-      });
-    }
-    
-    if (error.name === 'JsonWebTokenError') {
-      return res.status(401).json({
-        success: false,
-        message: 'Invalid token'
-      });
-    }
-
-    return res.status(500).json({
-      success: false,
-      message: 'Authentication failed'
-    });
-  }
-};
-
-// Optional Authentication (for routes that work with or without auth)
-const optionalAuth = async (req, res, next) => {
-  try {
-    const authHeader = req.headers.authorization;
-    const token = authHeader && authHeader.split(' ')[1];
-    
-    if (token) {
-      const decoded = verifyToken(token);
-      const user = await User.findById(decoded.userId).select('-password -refreshTokens');
-      
-      if (user && user.isActive) {
-        req.user = user;
-        req.userId = user._id;
-      }
-    }
-    
-    next();
-  } catch (error) {
-    // Continue without authentication if token is invalid
-    next();
-  }
-};
-
-// Role-based authorization
-const requireRole = (roles) => {
-  return (req, res, next) => {
-    if (!req.user) {
-      return res.status(401).json({
-        success: false,
-        message: 'Authentication required'
-      });
-    }
-
-    if (!roles.includes(req.user.role)) {
-      return res.status(403).json({
-        success: false,
-        message: 'Insufficient permissions'
-      });
-    }
-
-    next();
-  };
-};
-
-// Session Management
-const sessionManager = {
-  activeSessions: new Map(),
-
-  createSession: (userId, deviceInfo) => {
-    const sessionId = `${userId}_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
-    this.activeSessions.set(sessionId, {
-      userId,
-      deviceInfo,
-      createdAt: new Date(),
-      lastActive: new Date()
-    });
-    return sessionId;
-  },
-
-  updateActivity: (sessionId) => {
-    const session = this.activeSessions.get(sessionId);
-    if (session) {
-      session.lastActive = new Date();
-    }
-  },
-
-  removeSession: (sessionId) => {
-    this.activeSessions.delete(sessionId);
-  },
-
-  getUserSessions: (userId) => {
-    const sessions = [];
-    for (const [sessionId, session] of this.activeSessions) {
-      if (session.userId.toString() === userId.toString()) {
-        sessions.push({ sessionId, ...session });
-      }
-    }
-    return sessions;
-  },
-
-  clearUserSessions: (userId) => {
-    for (const [sessionId, session] of this.activeSessions) {
-      if (session.userId.toString() === userId.toString()) {
-        this.activeSessions.delete(sessionId);
-      }
-    }
-  }
-};
-
-module.exports = {
-  generateAccessToken,
-  generateRefreshToken,
-  verifyToken,
-  authenticateToken,
-  optionalAuth,
-  requireRole,
-  sessionManager
-};
