+++ backend/routes/sessionRoutes.js
@@ -1,11 +1,11 @@
 import express from 'express';
 import { validateSession, logout } from '../controllers/authController.js';
-import { verifyToken } from '../middleware/authMiddleware.js';
+import { authenticateToken } from '../middleware/authMiddleware.js';
 
 const router = express.Router();
 
 // Session management routes
-router.get('/validate', verifyToken, validateSession);
-router.post('/logout', verifyToken, logout);
+router.get('/validate', authenticateToken, validateSession);
+router.post('/logout', authenticateToken, logout);
 
 export default router;
